Update UI with footer and better buttons
New bug to fix, it keeps the winner even after you start a new 5 round game. Might
need to add either a new game button after the first 5 rounds are done, or rework the
logic of how it displays the winner.

There is an option to add a modal for the winner, that way once its closed it doesn't
matter if it doesn't disappear as it will be hidden.